formatting and minor fixes Explicitly init uint as uint256

+++ contracts/mockups/RealitioMockup.sol
@@ -3,25 +3,29 @@ pragma solidity ^0.7.5;
 
 import "hardhat/console.sol";
 
-// this is only for ganache testing. Public chain deployments will use the existing Realitio contracts. 
+// this is only for ganache testing. Public chain deployments will use the existing Realitio contracts.
 
-contract RealitioMockup
-
-{
-    uint result = 420;
+contract RealitioMockup {
+    uint256 result = 420;
     bytes32 public actualQuestionId;
     bytes32 public actualContentHash;
 
-    function setResult(uint _result) public
-    {
+    function setResult(uint256 _result) public {
         if (_result == 69) {
-            result = ((2**256)-1);
+            result = ((2**256) - 1);
         } else {
             result = _result;
         }
     }
 
-    function askQuestion(uint256 template_id, string calldata question, address arbitrator, uint32 timeout, uint32 opening_ts, uint256 nonce) external payable returns (bytes32) {
+    function askQuestion(
+        uint256 template_id,
+        string calldata question,
+        address arbitrator,
+        uint32 timeout,
+        uint32 opening_ts,
+        uint256 nonce
+    ) external payable returns (bytes32) {
         bytes32 content_hash = keccak256(abi.encodePacked(template_id, opening_ts, question));
         bytes32 question_id = keccak256(abi.encodePacked(content_hash, arbitrator, timeout, msg.sender, nonce));
         actualContentHash = content_hash;
@@ -54,6 +58,4 @@ contract RealitioMockup
             return 0;
         }
     }
-
 }
-
