+++ contracts/mockups/redeploys/BridgeMockupV2.sol
@@ -3,24 +3,27 @@ pragma solidity ^0.7.5;
 
 import "hardhat/console.sol";
 
-import '../../interfaces/IRCProxyXdai.sol';
-import '../../interfaces/IRCProxyMainnet.sol';
+import "../../interfaces/IRCProxyXdai.sol";
+import "../../interfaces/IRCProxyMainnet.sol";
 
 // a mockup to test changing the proxy, this is as per the original has a new number variable which is read
-contract BridgeMockupV2
-{
+contract BridgeMockupV2 {
     address public oracleProxyMainnetAddress;
     address public oracleProxyXdaiAddress;
-    uint public number;
+    uint256 public number;
 
-    function requireToPassMessage(address _RCProxyAddress, bytes calldata _data, uint256 _gasLimit) external {
+    function requireToPassMessage(
+        address _RCProxyAddress,
+        bytes calldata _data,
+        uint256 _gasLimit
+    ) external {
         _gasLimit;
         _RCProxyAddress;
         _data;
         number = 69;
     }
 
-    function messageSender() external view returns(address)  {
+    function messageSender() external view returns (address) {
         if (msg.sender == oracleProxyMainnetAddress) {
             return oracleProxyXdaiAddress;
         } else {
@@ -35,7 +38,4 @@ contract BridgeMockupV2
     function setProxyXdaiAddress(address _newAddress) external {
         oracleProxyXdaiAddress = _newAddress;
     }
-
-
 }
-
